:sparkles: (CardIcon) extract component from DetailsCard
