+++ frontend/src/components/CardIcon/index.tsx
@@ -1,0 +1,1 @@
+export { CardIcon } from './CardIcon';
